refactor wallet model

+++ backend/app/models/wallet.py
@@ -3,4 +3,4 @@ from beanie import Document
 
 class Wallet(Document):
     nickname: str
-    balance: int = 0
+    balance: float = 0
